Update patch set 4
Patch Set 4: Verified+1 Trybot-Ready+1

(1 comment)

I believe this is okay to go in now.  We can revisit and potentially lessen the requirement for restrictions in the manifest in the future once vendeb if we decide that is the correct behaviour.

Patch-set: 4
Reviewer: Gerrit User 1108834 <1108834@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=+1
Label: Trybot-Ready=+1