Update patch set 1
Patch Set 1:

I don't see a big problem with this change other than that it might lead to a fight between update_engine and vboot that could wear NVRAM down. I don't think it makes sense to have a max_rollforward value smaller than the current TPM value, so updating it shouldn't be a problem, but we would have to make sure update_engine knows about this and never tries to set the value below the current TPM value even if policy tells it to. I think it's easier to just leave this value fully controlled by update_engine only.

However, I also don't think it fixes the issue from crbug.com/900002: that problem only exists when your actual rollback version in the TPM is also 0, so this wouldn't update anything. Are you submitting this to solve a different problem? If so, please describe the issue you're seeing and how this fixes it.

Patch-set: 1
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>