Update patch set 2
Patch Set 2: Code-Review-1

This feels like the wrong fix to the problem.

Zero is valid (meaning "never roll forward past where we are now").  I don't see a problem with how it's used in vboot_api_kernel.c.  This is the only place in vboot_reference that depends on that value, and storing it back to NvStorage won't affect how that behaves.

Can you point to what else depends on the value of kernel_max_rollforward?  That is, why does this fix actually fix the problem?

Patch-set: 2
Reviewer: Gerrit User 1000187 <1000187@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1