Update patch set 2
Patch Set 2:

@Randall: Please have a look at crbug.com/900002 and let me know what you think. I think the issue that's blocking FAFT here is in update_engine, but the way we report TPM kernel versions in crossystem is still a bit weird right now (and, in particular, awkwardly inconsistent between normal and dev mode), so I think we may still want to fix something there.

> We were seeing failure in below FAFT tests
> 
> •	firmware_TPMVersionCheck.dev
> •	firmware_TPMVersionCheck
> •	firmware_TPMVersionCheck.ec_wp
> 
> After further debug, we have realized that Kernel_max_rollback always set to 0 which leads tpmker_ver not updating with correct tpm_firmware_version. 
> 
> Later, we have figure out that, if we made Kernel_max_rollback > 0 then tpmker_ver has valid value (same as tpm_fimware_version)

But I don't think this patch really fixes that problem for you. Maybe you accidentally fixed it independently by rolling your TPM version forward and then thought that it was this patch that fixed it, because once the TPM version for a device has rolled forward you won't see this problem anymore (even with the original code). But I don't see how this patch fixes anything.

If the rollback version in the TPM is already 0x10001, then it doesn't matter what your kernel_max_rollforward is set to... due to the line right above the one you're changing, vboot will interpret it as if it were equal to the version in the TPM (even if it doesn't write that back). crossystem tpm_kernver will return 0x10001 and firmware_TPMVersionCheck will pass.

The only way firmware_TPMVersionCheck can fail here is if the version stored in the TPM is still 0x0, not 0x10001. In that case, if kernel_max_rollforward is also 0x0, the max_rollforward feature will work as intended and prevent the version in the TPM from rolling forward to 0x10001. You patch wouldn't change anything about that, because both versions are still 0x0 at that time. The real bug is in update_engine which is supposed to always set kernel_max_rollforward to 0xfffffffe for non-enrolled devices, but doesn't seem to do that in all cases.

Patch-set: 2
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>