Update patch set 1
Patch Set 1: Code-Review-1

I'm not convinced this is the right way to approach this, even if we want this at all. In general, I'm always worried that ignoring errors will make us miss important problems during testing and dogfooding. At the very least, the string you log should start with "ERROR: " because some tests check for that.

Also, if we do want this I think it should be enabled in the depthcharge part, not here. Right now, depthcharge immediately aborts trying to update anything when one of the TCPCs fails. This makes sense if we're going straight into recovery, but if we want to implement a sort of "best effort" update scheme, we should make sure it will still attempt to update the remaining TCPCs even if one of them failed somewhere.

Patch-set: 1
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1