Update patch set 2
Patch Set 2: Code-Review-1

This sounds like a specific hack for Eve that will not be necessary in Campfire 2.0 (because then the EC-RO will support Alt OS stuff). So why are we cherry-picking this into ToT? If you need to have support for the NVRAM flag in crossystem that is one thing, but then you should exclude the firmware parts from this patch.

Patch-set: 2
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1