Update patch set 2
Patch Set 2:

Given Julius's analysis, I'm convinced max_rollforward is working as intended.  The problem is that in developer mode, crossystem tpm_kernver is a hybrid of (the version in the TPM at the end) and (the version we actually booted).  That confuses the test.

Patch-set: 2
Tag: mailMessageId=<CAALoFELO9scOtUxE2Y7nd1pLByGTjfbGGukbHU+aZgm9JO472w@mail.gmail.com>