Update patch set 5
Patch Set 5:

> For me, in developer mode when GBB flag is set to 0x39, i'm seeing tpmkern_val value is set to 0x10001 same as fw tpm version. but the real problem is when we are in normal mode, then fw version of tpm is 0x10001 but tpmkern_val is 0. in both cases kernel_max_rollback value is 0.
> 
> After further debugging, it appears to me that, during vboot_normal function, below code has force to make tpmker_val change to 0 based on max_rollback check.

Sorry, but I don't see how this can be the case. This is the current code:

	if (max_rollforward < shared->kernel_version_tpm_start)
		max_rollforward = shared->kernel_version_tpm_start;

	if (shared->kernel_version_tpm > max_rollforward) {
		VB2_DEBUG("Limiting TPM kernel version roll-forward "
			  "to 0x%x < 0x%x\n",
			  max_rollforward, shared->kernel_version_tpm);

		shared->kernel_version_tpm = max_rollforward;
	}

If, as you say, the version in the TPM is actually 0x10001, then shared->kernel_version_tpm_start will be 0x10001. That means that the first if() will trigger (because 0 < 0x10001), and max_rollforward will also be set to 0x10001. So shared->kernel_version_tpm will not actually be greater than max_rollforward (and even if that path got executed, it would still just be set to 0x10001).

The only way I can really see a problem here is if shared->kernel_version_tpm_start (meaning the version actually in the TPM) is 0x0. Are you *absolutely sure* that the version in the TPM on the device you're testing this on is 0x10001? Note that 'crossystem tpm_kernver' is not necessarily returning the right value for this (in developer mode it returns something else). If you want to read the raw TPM space, you can run this from userspace:

  stop trunksd
  tpmc read 0x1008

Patch-set: 5
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>