Update patch set 5
Patch Set 5:

> Yes, i'm sure that, i have seen the cases where shared->kernel_version_tpm is 0x10001 but shared->kernel_version_tpm_start is 0 hence it don't enter into first if condition but it do enter into 2nd if case where it again limits kernel_version_tpm to 0 as max_rollback itself is 0 at that point.

But shared->kernel_version_tpm_start *is* the version number currently in the TPM! Or do you see any way it could not be? shared->kernel_version_tpm is not necessarily the version in the TPM, after the kernel has been loaded it contains the version of the kernel that was loaded if that is higher. I know the naming isn't great, but that's how that code works.

> >  stop trunksd
> >   tpmc read 0x1008 4
> 
> i'm read something like this
> 
> 2 4c 57 52

That's just the struct version and a signature value, these are constant and always there. The actual rollback version comes later (bytes 5-8). This is how it looks for 0x10001:

> localhost ~ # tpmc read 0x1008 0xd
> 2 4c 57 52 47 1 0 1 0 0 0 0 55

...and this is how it would look for 0x0:

localhost ~ # tpmc read 0x1008 0xd
2 4c 57 52 47 0 0 0 0 0 0 0 e8

Patch-set: 5
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>