vboot_reference: add flag to record booting through Alt OS legacy mode
When booting into Alt OS legacy mode, we plan to disable TPM before
handing off control to the OS.  On S3 resume, the TPM is re-enabled,
and thus we need to disable it in coreboot's S3 resume path.
coreboot needs to know whether or not the system is currently booted
into Alt OS legacy mode, so that it only disables TPM in this case.

When the decision is made to boot through Alt OS legacy mode, record
in a VBSD flag in case this information is needed later on.

Additionally, create a VB2_SD version of the same flag (used by
coreboot).

BUG=b:70681930,b:118202153,b:118787888
TEST=None

Change-Id: I95005280d507e78b67c30c4cee442ae86c9d0592