Update patch set 5
Patch Set 5:

> but due to the fact that, there is 2nd if case which limits shared->kernel_version_tpm to Zero based on max_rollforward. I'm unable to understand why entire log has made around max_rollback and there is no way today that max_rollback can be uprev in code.

max_rollforward isn't supposed to be zero, that's the whole bug here. Please follow crbug.com/900002 for details. update_engine is supposed to set max_rollforward to 0xfffffffe, and it (hopefully?) does that on production devices, but it seems to currently not do that for early bring-up devices (I'm not quite sure what the differentiating factor is, it might just not work for test images in general). This is a bug in update_engine that update_engine people need to fix.

Patch-set: 5
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>