futility: updater: Revise output messages for debug logs
In auto update and recovery, the firmware updater was executed with both
stdout and stderr logged. However, the logs usually comes with all stderr
first then all stdout. This makes it harder to debug because the
messages logged in out of order.

TO solve that, few macros are introduced:
 INFO: for useful information.
 STATUS: the most common information, usually comes with a prefix code.

And all messages should now go to stderr except the final execution
result (and those output commands, for example --manifest).

BUG=chromium:875551
TEST=TEST=make futil; tests/futility/run_test_scripts.sh $(pwd)/build/futility
BRANCH=None

Change-Id: Ie0dc6594ece10e7e15caf9c36353e2b3ec8754c5
Signed-off-by: Hung-Te Lin <hungte@chromium.org>