Allow firmware library to be built without stdint.h
While stdint.h is the standard C99 way of getting access to types like
uint32_t and uint8_t, for more kernel-like bootloaders such as U-Boot,
stdint is not used. Instead the types are defined locally to match the
toolchain in use, such that (for example) "%lld" can be used with uint64_t
in all cases.

In fact, the kernel uses u8 and u32, but these are aliased to their more
long-winded cousins.

Allow stdint.h to be optional, under control of NO_STDINT, passed in from
the bootloader.

BUG=chromium:908228
TEST=FEATURES=test emerge-grunt vboot_reference

Change-Id: I32db7961ee5812fc1078d5514e879ff9d37eb588
Signed-off-by: Simon Glass <sjg@chromium.org>