WIP: vboot_api_kernel: check TPM mode on normal boot
When booting into Alt OS legacy mode, we plan to disable TPM
before handing off control to the OS.  On a warm reboot back
to Chrome OS, we must check the TPM mode.  If it is disabled,
a hard reboot should be triggered to restore TPM functionality.

Add this check to VbBootNormal.  Only accept the TPM mode
TpmModeEnabledTentative (0).

BUG=b:119203340
TEST=compile, flash, and boot eve

Change-Id: If6097aa67eb135e24c735bd7948ec25091bed3be