Update patch set 8
Patch Set 8:

what is generating this json-but-not-json file ?  we should fix it to actually be json.  these hand edits are very fragile which is the sort of thing you don't want in code responsible for signing the secure element.

comments are usually easy to filter out (`sed -i '/^ *#/d'`), but hex constants are not.

Patch-set: 8
Reviewer: Gerrit User 1000768 <1000768@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>