Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> I just want to get this off my plate so I can get back to what I'm supposed to be working on without the constant context switching.
> 
> I converted a JSON  file by stripping out the # comments and manually converting the 0x hex values to decimal and it seems to work and result in the same signed image.  Unless there are concrete objections of something that will not work with these changes, I'm planning to modify the scripts to use jq or python (vapier@: do you have a preference? jq isn't on the signers) and then cr50 team fix up their JSON files in the various branches they want signed.  There will need to be at least one build made with these changes to get suitable test input data for the cros-signer unit tests to pass.
> 
> Eg of converted hex values:
>     "board_id": 52,
>     "board_id_flags": 16,
>     "board_id_mask": 0,

I don't think this is how it is supposed to work: we need to be able to keep building signed cr50 images locally, so the builder and local repos need to stay compatible at all times.

Also, not being able to use hex values is quite a pain, this alone is a good reason to think twice before making changes to the existing flow.

I would suggest to open a bug to migrate signer/builder to 'proper' JSON and get the current flow working first.

Patch-set: 8