Update patch set 8
Patch Set 8:

there is valid JSON and there is invalid JSON.  tooling is pretty consistent along those lines.  and while i hate the fact JSON doesn't support hex values (while simultaneously supporting a ridiculous plethora of scientific notations), that's how it is.

i'm not saying the cr50 tooling itself can't support whatever extensions it wants, but if it's going to try to export these files for other tools to process, it needs to be in valid JSON format.

wrt the cr50-utils repo, i see code that reads these config files (codesigner.cc), but i'm not seeing what code actually generates the files which is what we'd want to change.

we can use `jq` on the signers.  just add it to the cros-signing/ repo as one of the required packages (and we can hand deploy it on existing installs).  jq should be easier to use than python in these shell scripts.

Patch-set: 8
Reviewer: Gerrit User 1000768 <1000768@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>