image_signing: drop support for old_verity_argv images
We changed the verity kernel command line form before R16 was released
and included backwards compat support in the scripts for it.  But all
the devices that were released for these old versions are EOL, and we
don't need to sign images that old anymore, so drop support.

BRANCH=None
BUG=chromium:891015
TEST=precq passes

Change-Id: I5cc37fae19fb4b3db229598aa0f5c69a6f32005a