Update patch set 1
Patch Set 1:

> I have created a small hashing library (for measured and verified boot). This library is used in bootblock/ramstage and romstage. This library uses the firmware/2lib/2rsa.c. The coreboot (default) build uses -Wmissing-prototypes, which causes the problem. I've tested with adding -Wno-error=missing-prototypes, but this failed for bootblock, ROMCC does not accept this parameter.

Sorry, I'm still somewhat confused what you're trying to do... maybe you could upload the coreboot code you have to build this to review.coreboot.org to make it clearer?

If you're somehow trying to directly build 2rsa.c as part of a coreboot stage, then that sounds like the wrong thing to do. vboot is a separate library and should remain that way, you shouldn't start picking individual pieces out of it. coreboot already has all the stuff in its build system to build and link vboot as a library if CONFIG_VBOOT is enabled... with this, you should be able to just call functions from it within coreboot code (but it is still built as a separate library with its own CFLAGS).

If you want to use these functions but don't want to set CONFIG_VBOOT, we should change the coreboot Makefiles to decouple that (so that the vboot_fw20.a library is still being built and available during linking even if we're not enabling all the coreboot verification code paths).

Patch-set: 1