Update patch set 3
Patch Set 3:

(1 comment)

There's one part that I thought is a little bit weird, but I'm not familiar with this code base and the GBB stuff so I may very well be wrong.

As for the calculation of digest, yes, it's correct to the best of my knowledge.

Patch-set: 3
Reviewer: Gerrit User 1314760 <1314760@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>