Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> It's too late in the project to update the fmap (Grunt and Octopus systems specifically).
> 
> Why would the update need to wipe the MRC cache?  Doing this will ALWAYS prevent the system from resuming from S3 if it's in use.  A reboot after resuming in this case is the BEST case scenario.
> 
> This is only an issue if the system is not in RO (pre-PVT units), so all of our dogfooders are getting system hangs and reboots.  It seems like we'd want the MRC cache to be preserved on those systems similar to what we see in PVT/MP units.

If we are doing a full RO update that could include moving/resizing the MRC region.

It could also include a new incompatible FSP that would have issues with the passed in stale data.  (FSP is supposed to be able to detect this, but I'm not convinced)

Patch-set: 1
CC: Gerrit User 1000423 <1000423@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>