Update patch set 1
Patch Set 1:

Sorry, I didn't pay attention to any of the altos stuff while it was branched, but if you now wanna merge it... can you explain how the boot flow works that could lead to the TPM being disabled at this point? A disabled Cr50 means it won't accept any TPM commands, right? We're using the TPM during firmware verification already, so if it was disabled since boot we should never even be able to get here.

Patch-set: 1
CC: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>