Update patch set 1
Patch Set 1:

> The expected flow to boot an alternate OS will be:
> 1. Cr50 boots, set TPM state to TPM_ENABLED_TENTATIVE
> 2. depthcharge checks TPM state - sees TPM_ENABLED_TENTATIVE and continues boot
> 2. User enters recovery screen or dev mode (TBD)
> 3. Capture TBD key-sequence to launch alt OS
> 4. Verify alt OS
> 6. Disable TPM on Cr50 - Cr50 disables I2C/SPI interface and key ladder. TPM state set to TPM_DISABLED.
> 7. Boot into alt OS
> 8. On AP reboot, the Cr50 automatically re-enables the I2C/SPI interface, but leaves the key ladder disabled.  TPM state set to TPM_ENABLED.
> 9. depthcharge checks TPM state.  State is not TPM_ENABLED_TENTATIVE, depthcharge issues IMMEDIATE_RESET to Cr50.
> 10. Cr50 boots (goto step 1)

Thanks, that makes more sense then. It seems that the TPM mode names are pretty misleading and should probably be renamed -- e.g. TPM_ENABLED_TENTATIVE is really more of a TPM_ENABLED, and TPM_ENABLED is more of a TPM_NEEDS_RESET.

Is there any reason why all of this is implemented in depthcharge? I understand that that may have been due to not being able to modify RO code on the branch, but that justification shouldn't count for master (the whole idea of branching was that we could clean all those things up on merge). Conceptionally, it feels to me like this would fit better into coreboot as part of tpm_setup(). (In coreboot it's also easier to distinguish cr50 from other TPMs, which you'd still need to add to this code anyway, since we have that info available as a compile-time config.)

Patch-set: 1