Update patch set 1
Patch Set 1:

> Regarding the TPM mode names, these came from the Cr50 firmware source, but I agree it makes sense to rename then in the AP firmware to match the actions needed.  The enum on the AP side could include comments indicating the Cr50 name for clarity.

Why not change the name in Cr50 code as well? ;) Changing an enum name doesn't break anything so you wouldn't need to make sure they get changed at the same time or anything... as long as it will get cleaned up everywhere at some point, that should be good enough.

> I appreciate the feedback - I'm brand new to working the coreboot and associated codebases, so I'm still learning how the pieces fit together.  I need to do some digging into the coreboot code to see where it might fit to perform the TPM commands noted in steps 2, 6, 9, and 10.

Note that step 6 probably needs to remain in depthcharge, since that is triggered by the UI (which is in depthcharge). But 2 and 9 (these are essentially the same check, right?) would fit better in coreboot, I think. In fact, we are already doing something very similar (asking Cr50 whether we should reset it) in src/vendorcode/google/chromeos/cr50_enable_update.c anyway... probably easiest to just tack it on there. (Actually, it would be great if we could check for both updates and the TPM mode in a single command to avoid having the TPM transaction boot time overhead twice... maybe the Cr50 team would be open to such a change?)

Patch-set: 1