Update patch set 1
Patch Set 1:

First off, I want address item #8 in the "expected flow" that Keith posted.  When an AP reset occurs, TPM2 BSS memory is cleared, and Cr50's tpm_mode value is reset to TPM_ENABLED_TENTATIVE.  However, since the key ladder is not available, the GET_MODE command actually returns VENDOR_RC_INTERNAL_ERROR (see board/cr50/tpm2/tpm_mode.c).

Second, re. naming of TPM_*:
There was actually a long discussion about whether or not it makes sense to lock the TPM in "enabled" mode.  The end-result was that it is superfluous (provides no security benefit), and thus currently only TPM_ENABLED_TENTATIVE and TPM_DISABLED were used.
So actually, these names *mostly* make sense.  It could also be simpler just to rename them to ENABLED and DISABLED.  But that would require removing one of the values, which could get more complicated than just renaming them (esp. since TENTATIVE = 0, ENABLED = 1, and DISABLED = 2).

Second, re. moving this code to coreboot:
(Actually, unless I'm mistaken, the enable_update function is run in ramstage, so there is no concern about RO/RW.)

It's really a trade-off of keeping all vboot-related code centralized in one location, versus spreading it out to earlier-on stages in coreboot.  The latter gives us faster boots, since we are detecting that we need a reset earlier on.  In this case, where this is a very seldom performed task, the time delta may not be an important concern.

If there is another good reason to move it to coreboot (e.g. logically fits better with tpm_setup), another consideration is pushing vboot_reference in the direction of becoming a useful "library" of vboot-related tasks, and calling that from coreboot, rather than relocating full blocks of code from depthcharge/vboot_reference into coreboot.

For example -- in this case, it would be nice to see the logic of "What is the mode of Cr50?  Is this an acceptable mode?  If not, issue a Cr50 reset request." within vboot_reference, with a call from coreboot into that code.

Julius, what are your thoughts?

Patch-set: 1
Reviewer: Gerrit User 1126709 <1126709@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>