Update patch set 1
Patch Set 1:

> First off, I want address item #8 in the "expected flow" that Keith posted.  When an AP reset occurs, TPM2 BSS memory is cleared, and Cr50's tpm_mode value is reset to TPM_ENABLED_TENTATIVE.  However, since the key ladder is not available, the GET_MODE command actually returns VENDOR_RC_INTERNAL_ERROR (see board/cr50/tpm2/tpm_mode.c).

So this patch will actually print "TPM encountered some error; reset Cr50" in the "normal" AltOS-to-CrOS switch case, not "Invalid TPM mode (%d, expected: %d); reset Cr50"? That's bad, then, the code should not be misleading like that. (Unless the plan was to fix this on the Cr50 side at some point, but then we should track that that actually happens.) In general, relying on some "internal error" code to denote a specific, valid state sounds like bad design... errors should be errors and valid states that are expected during normal execution should not (or at least they should have a separate error code which makes that clear).

Also, why does GET_MODE not work when the key ladder is disabled? It doesn't really access any data secured by that, right? Which commands do work in this state, exactly, and which don't? 

> Second, re. naming of TPM_*:
> There was actually a long discussion about whether or not it makes sense to lock the TPM in "enabled" mode.  The end-result was that it is superfluous (provides no security benefit), and thus currently only TPM_ENABLED_TENTATIVE and TPM_DISABLED were used.
> So actually, these names *mostly* make sense.  It could also be simpler just to rename them to ENABLED and DISABLED.  But that would require removing one of the values, which could get more complicated than just renaming them (esp. since TENTATIVE = 0, ENABLED = 1, and DISABLED = 2).

Just rename them to ENABLED = 0, DEPRECATED_DO_NOT_USE = 1, DISABLED = 2 in the enum (on both sides)? 

> 
> Second, re. moving this code to coreboot:
> (Actually, unless I'm mistaken, the enable_update function is run in ramstage, so there is no concern about RO/RW.)

Right, I first thought about putting it in tpm_setup() (which is in RO), but later realized that we're already doing the update check thing later anyway so it might fit well with that.

> It's really a trade-off of keeping all vboot-related code centralized in one location, versus spreading it out to earlier-on stages in coreboot.  The latter gives us faster boots, since we are detecting that we need a reset earlier on.  In this case, where this is a very seldom performed task, the time delta may not be an important concern.
> 
> If there is another good reason to move it to coreboot (e.g. logically fits better with tpm_setup), another consideration is pushing vboot_reference in the direction of becoming a useful "library" of vboot-related tasks, and calling that from coreboot, rather than relocating full blocks of code from depthcharge/vboot_reference into coreboot.

Well... I don't think this really belongs in vboot at all, to be honest. It's getting introduced as part of the AltOS feature, yeah, but it's really just about handling a specific Cr50 state. Until now, all code handling specific Cr50 details (at both the coreboot and depthcharge stages of boot) has been kept outside of vboot. The more I think about it, the more I feel that this really is conceptually very similar to the update check and it makes perfect sense to put them both together in a single "check whether Cr50 wants any special treatment this boot" step. And neither of them really have anything to do with verification. (In fact, even AltOS doesn't really have much to do with vboot, it's just mostly implemented there because vboot holds the UI monopoly at the moment.)

For the boot speed side, I think normal mode Chrome OS boot is the most important thing, with the AltOS-to-CrOS transition probably much rarer and a secondary concern at most (especially because what you're landing here affects all boards, not just boards that really support AltOS). Every extra TPM access takes a bit of time (though not that much, but it adds up) so it's nice to try to minimize them where possible in the normal boot path. That's why I'd really like to see this merged with the update check if possible, I think it's a good fit conceptually and will save us another round trip (this could be as simple as making CMD_TURN_UPDATE_ON return a VENDOR_RC_NEEDS_REBOOT error or something like that).

Patch-set: 1