Update patch set 1
Patch Set 1: Code-Review+1

Thanks for improving the messages. One thing that confused me when I was first starting was the generic term "write protection". For the FPMCU, we use the hardware write protection (GPIO) and have an additional software write protection in the firmware. I'm not sure if your firmware has both types, but if it does, it would be good to specify which one needs to be disabled. Maybe even the error message could have a go link with instructions on how to disable? (Though maybe that is too hard to keep up to date)

Patch-set: 1
Reviewer: Gerrit User 1314845 <1314845@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1