Update patch set 2
Patch Set 2:

> Eck, you're right.  I'm tempted just to chump the CL since it's so simple.  But for the sake of continuity, I suppose I will do it the "right way".  I wonder how long we have to wait for vboot submodule to uprev?

Uhh... no, please don't chump it, that wouldn't help. You'd still need to uprev the submodule upstream (and then get your CL submitted, and cherry-pick it to Chromium, and chump it there) anyway, and the tree would be broken in the meantime. If you want a different option, you could make an exception to our usual "upstream first" rule and submit your coreboot CL to Chromium first... then submit this, then submit your coreboot changes upstream together with an uprev in the same CL. For a real circular dependency that would probably be the way to go, but for this, I think add-first-remove-later is the easiest option.

You don't need to wait for anyone to uprev the submodule upstream, just do it yourself. We don't have an official uprev cadence, it's just done as needed by whoever needs a new vboot version.

Patch-set: 2
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>