vboot: add power button query/disable functionality
The primary purpose of this change is to add a flag to VbExGetSwitches()
so the current press/depress status of the power button can be queried.
However, on x86 systems the system will just power off, so this change
also adds a VbExEcEnablePowerButton() that can be used to disable the
power button from turning off the system (this disable is intended to be
temporary--only for as long as the power button is expected to be
presses).  I would have likes to use "Disable" rather than "Enable" in
the name, but I wanted to stay consistent with the naming scheme that's
used within depthcharge.

BUG=b:124358784
BRANCH=none
TEST=Locally built and flashed. Confirmed that the power button status
can be queried when the power button is disabled.
CQ-DEPEND=CL:1470273

Change-Id: I579ebe657ae35fb866eb30b466e8e8c16f54e584
Signed-off-by: Matt Delco <delco@google.com>