vboot: add diagnostic mode
This change adds diagnostic mode.  When enabled for a board (based on
defconfig in depthcharge) the user can press Ctrl-C or F12 at a recovery
mode screen, at which point the user is asked to confirm by pressing the
power button (Esc or 30 second timeout returns to prior screen).  If
confirmed, a nv bit is set and the system reboots.  Upon reboot, if the
nv bit is set then the user is again prompted to confirm via power
button.  If user confirms then diagnostic payload is verified and run.
Otherwise a recovery reason is recorded and system reboots to recovery
mode.

BUG=b:124358784
BRANCH=None
TEST=Locally built and flashed using change that enabled for atlas and
set to use payload 2 (tianocore) rather than 5 (diagnostic).  Confirmed
that Ctrl-C is functionality or not based on defconfig.  Confirmed that
Esc or timeout returns to recovery screen.  Confirmed that pressing power
button sets NV bit and reboots.  Confirmed that NV bit can be set and
queried via crossystem. Confirmed that during boot confirmation screen
appears or not based on NV bit.  Confirmed that pressing power button
caused payload to be verified and run.  Confirmed that non-matching hash
(build configured to use sha1 rather than sha256) caused payload to not
be run and system reboot to recovery.  Confirmed that Esc or timeout
caused system to resume with normal boot (e.g., dev menu).

Change-Id: I8979d4eeb443bf64b727ee86a814c46d1d27ff37
Signed-off-by: Matt Delco <delco@google.com>