Update patch set 11
Patch Set 11:

(3 comments)

> How does this look instead, it gets rid of the ifdefs and I think the optimizer will be able to compile out the vendor_data_ui functions when VENDOR_DATA_LENGTH is 0

I think this works, but it depends on a bit of weirdness on the depthcharge side in how Kconfig handles defaults and dependencies. I think it may be easier to follow if we just get rid of CONFIG_VENDOR_DATA_UI in depthcharge, set the default for CONFIG_VENDOR_DATA_LENGTH to 0, and then say that the option is the source of truth for both whether the feature is enabled and how long the string is across both vboot and depthcharge.

Patch-set: 11
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>